Typescriptify pages
