+++ components/Layout/index.tsx
@@ -1,0 +1,6 @@
+import ContentLayout from './ContentLayout'
+import DefaultLayout from './DefaultLayout'
+import Perks from './Perks'
+import Sidebar from './Sidebar'
+
+export { ContentLayout, DefaultLayout, Perks, Sidebar }
