--- pages/_app.js
@@ -1,6 +1,0 @@
-import '../styles/index.scss'
-
-// This default export is required in a new `pages/_app.js` file.
-export default function MyApp({ Component, pageProps }) {
-  return <Component {...pageProps} />
-}
