+++ pages/_app.tsx
@@ -1,0 +1,8 @@
+import { AppProps } from 'next/app'
+import '../styles/index.scss'
+
+function MyApp({ Component, pageProps }: AppProps) {
+  return <Component {...pageProps} />
+}
+
+export default MyApp
